[GTK] Expose API to report theme color changes https://bugs.webkit.org/show_bug.cgi?id=294389
Reviewed by Carlos Garcia Campos.

Enable the WebKitWebView:theme-color property for the GTK port, based on
the implementation added for WPE in 296035@main. This mainly involved
removing PLATFORM(WPE) guards or adding PLATFORM(GTK) ones in code
common for both ports, and keeping some of the guards to use GdkRGBA
instead of WebKitColor to represent colors in the public API. The
existing test case is enabled for the GTK port, using the ColorType
alias type (which maps to GdkRGBA or WebKitColor depending on the port).
A small webkitWebViewEmitThemeColorChanged() private helper function was
also added, then changed the WPE port to make both ports use it.

Canonical link: https://commits.webkit.org/296160@main