[JSC] Integer Range optimization should handle constant length https://bugs.webkit.org/show_bug.cgi?id=294467 rdar://problem/153329198
Reviewed by Yijia Huang.

This patch allows CheckInBounds(@index, $constant) to be analyzed in
integer range optimization. Previously we only allowed
CheckInBounds(@index, @value), and @value is typically length. But it is
possible that length can be constant folded. So we should handle
$constant case too. We also add ArithBitAnd rule which offers narrower
integer range as a result. This can be seen in some code places.

* Source/JavaScriptCore/dfg/DFGIntegerRangeOptimizationPhase.cpp:

Canonical link: https://commits.webkit.org/296213@main