Fix crash in WebFrameProxy::didCreateSubframe with site isolation on https://bugs.webkit.org/show_bug.cgi?id=294498
Reviewed by Charlie Wolfe.

Site isolation introduces a race condition that has caused several crashes for me today.
If a frame navigates from one site to another and when the frame load commits in the next
process, the previous process creates a new iframe, we can hit this assertion.  When this
condition is hit, it is correct to ignore the message from the previous process and do nothing.

* Source/WebKit/UIProcess/WebFrameProxy.cpp:
(WebKit::WebFrameProxy::didCreateSubframe):

Canonical link: https://commits.webkit.org/296234@main