WebGPU 1D texture sampling bug https://bugs.webkit.org/show_bug.cgi?id=294223 rdar://152880664
Reviewed by Tadeu Zagallo.

The docs for copyFromBuffer:...toTexture:...
do not allow for passing 0 to sourceBytesPerRow:
like they do for replaceRegion: calls. While it
seemed to work and is not detected by the Metal
validation layer, this appears to be a bug in WebKit
relying on undocumented behavior as it specifically
does not work in certain cases (sub buffer copies
from offsets not aligned to 64 bytes).

Test: api,operation,command_buffer,image_copy:* continues to pass, which
is the reason this zero was added.

Expectations are updated, all passing with this change.

* Source/WebGPU/WebGPU/Queue.mm:
(WebGPU::Queue::writeTexture):
(WebGPU::Queue::newTemporaryBufferWithBytes):

* LayoutTests/http/tests/webgpu/webgpu/api/operation/command_buffer/image_copy-expected.txt:
All sub-cases pass.

Canonical link: https://commits.webkit.org/296260@main