[WPE] Crash when running the mdn-bcd-collector test suite after 292932@main https://bugs.webkit.org/show_bug.cgi?id=293786
Reviewed by Xabier Rodriguez-Calvar.

A new DesktopPortal client is now created for each enumerateDevices() call, in order to prevent
recursion if the call site is not properly awaiting the result. The computeCaptureDevices completion
handler is also now chained from the PipeWireCaptureDeviceManager to the DesktopPortal, allowing us
to keep the openCameraPipewireRemote() method private to DesktopPortal.

Canonical link: https://commits.webkit.org/296261@main