Call owner document's checkCompleted in WebRemoteFrameClient::frameDetached https://bugs.webkit.org/show_bug.cgi?id=294569
Reviewed by Charlie Wolfe.

If we remove a RemoteFrame from the tree before or during the load event, then by the time
WebPage::didFinishLoadInAnotherProcess is called there is no more frame to be found so we
do nothing and the load event of the parent is never called.  To fix this, when removing a
LocalFrame, call the parent document's checkCompleted function if it is in the same process.
Document::checkCompleted is intended to be able to be called multiple times, and if it is
not completed loading yet it does nothing.

* LayoutTests/http/tests/site-isolation/resources/update-background.html:
* Source/WebCore/dom/Document.h:
* Source/WebKit/WebProcess/WebCoreSupport/WebRemoteFrameClient.cpp:
(WebKit::WebRemoteFrameClient::frameDetached):

Canonical link: https://commits.webkit.org/296294@main