[tables] ASSERT in WebCore::RenderTableSection::layoutRows (needsLayout) https://bugs.webkit.org/show_bug.cgi?id=294611 <rdar://153541166>
Reviewed by Antti Koivisto.

setNeedsLayout and setChildNeedsLayout have slightly different behavior when
a descendent marks itself dirty and we walk the containing block chain until
we find an dirty ancestor. This traversal does not stop when the ancestor
hash the self-needs-layout bit set (setNeedsLayout) only when set-normal-child-needs-layout
is set (setChildNeedsLayout).
Now during layout it's normally okay to mark the containing block chain dirty it is supposed to
be dirty. However in this case we mark the table cell from table section layout meaning
that the in-between table row is already cleared. If we mark it dirty at this point (through the ancestor chain marking), we finish layout with this renderer still needing layout.

* LayoutTests/fast/table/table-section-missing-content-expected.txt: Added.
* LayoutTests/fast/table/table-section-missing-content.html: Added.
* Source/WebCore/rendering/RenderTableCell.cpp:
(WebCore::RenderTableCell::computeIntrinsicPadding):
* Source/WebCore/rendering/RenderTableCell.h:
* Source/WebCore/rendering/RenderTableSection.cpp: replace setNeedsLayout with setChildNeedsLayout which is what this function normally calls.
(WebCore::RenderTableSection::layoutRows):

Canonical link: https://commits.webkit.org/296328@main