[visionOS] Guard external playback from spatial and immersive video https://bugs.webkit.org/show_bug.cgi?id=294336 rdar://147052015
Reviewed by Ada Chan.

For immersive video, the API should reject attempts to enter when the content is
detected as immersive.

For spatial media, ensure that LinearMediaPlayer does NOT use the peculiar since
it is not setup in external playback.

* Source/WebKit/Platform/ios/VideoPresentationInterfaceLMK.mm:
(WebKit::VideoPresentationInterfaceLMK::enterExternalPlayback):
* Source/WebKit/WebKitSwift/LinearMediaKit/LinearMediaPlayer.swift:
(WKSLinearMediaPlayer.isImmersiveVideo):
(WKSLinearMediaPlayer.makeDefaultEntity):

Canonical link: https://commits.webkit.org/296350@main