IntersectionObserver: Node leak test is flaky after 296279@main https://bugs.webkit.org/show_bug.cgi?id=294628
Reviewed by Simon Fraser.

296279@main made notification delivery asynchronous for JS intersection observers.
The test that would check for nodes being leaked by the observer was doing gc after an
rAF + zero delay timer. Since we keep pending notification targets alive until we deliver,
this effectively was causing a race between the EventLoop's task queue executing the notification
and the zero delay timer running. Sometimes we wouldn't deliver notifications before attempting a GC.

This change adds a small timeout between the rendering update where we run the requestAnimationFrame
callback and when we run gc to try to collect nodes so that the EventLoop's task queue has a chance to
execute.

* LayoutTests/intersection-observer/intersection-observer-should-not-leak-observed-nodes.html:

Canonical link: https://commits.webkit.org/296352@main