RELEASE_ASSERT fired in WebKit::ModelProcessModelPlayer::didUnload https://bugs.webkit.org/show_bug.cgi?id=294589 rdar://153047027
Reviewed by Mike Wyrzykowski.

This is a speculative fix as I can't reproduce this locally.
If this gets called when the model process connection is closed
and the page has already been closed, ModelProcessModelPlayer::modelProcessEnabled()
will return false because the page is no longer there. In that case,
we want to just return early and not fail the assertion.

* Source/WebKit/WebProcess/Model/ModelProcessModelPlayer.cpp:
(WebKit::ModelProcessModelPlayer::didUnload):

Canonical link: https://commits.webkit.org/296362@main