[Tools][WPE][GTK] Support running browserperfdash-benchmark inside the flatpak SDK https://bugs.webkit.org/show_bug.cgi?id=294723
Reviewed by Philippe Normand.

Currently browserperfdash-benchmark is running outside of Flatpak SDK and the browser
is running inside (via executing run-minibrowser script).

However, to support the upcoming plugin-plans (bug 294625) it is needed to run the runner
also inside Flatpak SDK because the the plugin process runs on the same pid than the runner.
And for this plugins to work on the right environment they should use the Flapak SDK one
and not the host one. For example, if they want to run 'ldd' that should happen inside the
flatpak SDK.

And this requires implementing a way of getting the TCP ports open from a given pid without
using 'lsof' or 'python-psutil' on simple_http_server_driver.py because neither of this two
tools is currently available on the Flatpak SDK.

* Tools/Scripts/browserperfdash-benchmark:
* Tools/Scripts/webkitpy/benchmark_runner/http_server_driver/simple_http_server_driver.py:
(linux_proc_get_listening_ports):
(SimpleHTTPServerDriver._find_http_server_port):

Canonical link: https://commits.webkit.org/296421@main