Web Inspector: REGRESSION STP Sources tabs fails to load when hitting a breakpoint rdar://153666768 https://bugs.webkit.org/show_bug.cgi?id=294627
Reviewed by Devin Rousso.

A SourceCodeTreeElement is not a ResourceTreeElement so it doesn't have
a `resource` property. This is probably what caused the second uncaught
exception in the bug report, "undefined is not an object (evaluating 'this.resource.sourceMaps')".

(The suspected regression point was https://github.com/WebKit/WebKit/pull/45103.)

* Source/WebInspectorUI/UserInterface/Views/SourceCodeTreeElement.js:
(WI.SourceCodeTreeElement.prototype.onpopulate):

Canonical link: https://commits.webkit.org/296422@main