List item marker should be rendered as the first child of the list item when `list-style-position: inside` is set https://bugs.webkit.org/show_bug.cgi?id=227000
Reviewed by Alan Baradlay.

As stated by https://developer.mozilla.org/en-US/docs/Web/CSS/list-style-position

"If a block element is the first child of a list element declared as `list-style-position: inside`,
then the block element is placed on the line after the marker-box."

In the current implementation, we render the marker on the same first line
as the content flow when the first child is a block element. This patch
changes the behavior to render the marker in the first line of the
list item, instead of the first line of the content when `list-style-position:inside` is present.

this align behavior with other browsers.
* LayoutTests/TestExpectations
* LayoutTests/fast/lists/list-marker-style-inside.html: Added.
* LayoutTests/fast/lists/list-marker-style-inside-expected.html: Added.
* LayoutTests/fast/repaint/list-marker-expected.txt: Added.
* LayoutTests/platform/mac/fast/repaint/list-marker-expected.txt: Removed.
* LayoutTests/platform/win/fast/repaint/list-marker-expected.txt: Removed.

* Source/WebCore/rendering/updating/RenderTreeBuilderList.cpp:
(WebCore::parentCandidateForMarker):

Canonical link: https://commits.webkit.org/296451@main