REGRESSION (293770@main): clicking on the "+" button in the Sources tab sidebar does nothing rdar://153193833 https://bugs.webkit.org/show_bug.cgi?id=294392
Reviewed by Patrick Angle.

If there are obscuredContentInsets applied to the WKInspectorWKWebView,
when InspectorFrontendHost is commanded to show a context menu, it
shouldn't base its location entirely on the MouseEvent that came from
the inspector's frontend. The frontend is unaware of the obscuredContentInsets
while the frame is. InspectorFrontendHost needs to apply that offset,
as otherwise the mouse event that the frame later creates to bring up
the context menu might happen on top of a wrong event target, one that
doesn't have a contextmenu event handler which would insert the expected
menu's contents.

Tested manually on macOS 26.0 where this bug also reproduced, while the
inspector having either a left inset (with the browser's sidebar open)
or a top inset (with the inspector undocked).

* Source/WebCore/inspector/InspectorFrontendHost.cpp:
(WebCore::InspectorFrontendHost::dispatchEventAsContextMenuEvent):

Canonical link: https://commits.webkit.org/296458@main